Refactor whitespace print
