+++ src/language-markdown/print/whitespace.js
@@ -1,5 +1,6 @@
 import { hardline, line, softline } from "../../document/index.js";
 import {
+  isNewLine,
   isSetextHeading,
   KIND_CJ_LETTER,
   KIND_CJK_PUNCTUATION,
@@ -265,4 +266,57 @@ function printWhitespace(path, value, proseWrap, isLink, options) {
   return canBeSpace ? " " : "";
 }
 
-export { printWhitespace };
+function printWhitespaceNode(path, options) {
+  return printWhitespace(
+    path,
+    path.node.value,
+    shouldPreventBreak(path, options) ? "never" : options.proseWrap,
+    false,
+    options,
+  );
+}
+
+function shouldPreventBreak(path, options) {
+  const { proseWrap } = options;
+  if (proseWrap === "never") {
+    return true;
+  }
+
+  if (
+    !path.isLast &&
+    // leading char that may cause different syntax
+    /^>|^(?:[*+-]|#{1,6}|\d+[).])$/.test(path.next.value) &&
+    // Avoid https://github.com/prettier/prettier/issues/18861
+    !hasFakeWhitespaceAfterNextToken(path) &&
+    // Next fake setext h2 `-` is going to be escaped, so no need to join adjacent words
+    !(proseWrap === "preserve" && isNextTokenFakeSetextH2Line(path))
+  ) {
+    return true;
+  }
+
+  return false;
+}
+
+function hasFakeWhitespaceAfterNextToken(path) {
+  const { siblings, index } = path;
+  if (!siblings || typeof index !== "number") {
+    return false;
+  }
+  const afterNext = siblings[index + 2];
+  return afterNext?.type === "whitespace" && afterNext.value === "";
+}
+
+/**
+ * @param {AstPath} path
+ * @returns {boolean}
+ */
+function isNextTokenFakeSetextH2Line(path) {
+  if (!isNewLine(path.node) || path.next?.value !== "-") {
+    return false;
+  }
+
+  const afterNext = path.siblings[path.index + 2];
+  return !afterNext || isNewLine(afterNext);
+}
+
+export { printWhitespace, printWhitespaceNode };
